pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cardex"
version = "0.1.1"
description = "Academic knowledge management system - from PDF to structured knowledge cards"
authors = [
    {name = "Justyn Chen", email = "your-email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "TBD"}
keywords = ["academic", "research", "pdf", "knowledge-management", "rag"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "streamlit>=1.30.0",
    "pymupdf>=1.23.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/matheme-justyn/cardex"
Documentation = "https://github.com/matheme-justyn/cardex/tree/main/docs"
Repository = "https://github.com/matheme-justyn/cardex"
Issues = "https://github.com/matheme-justyn/cardex/issues"

[project.scripts]
cardex = "cardex.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cardex*"]
exclude = ["tests*", "docs*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
